Assure event mask is set properly to get positions
The event mask can be retrieved in the context of
client thread as zero and then queued up to go to
msg task thread. By the time the msg is actually
handled in msg task thread, the actual event
mask at LOC API layer may have already changed, but
this mask would then be overridden by zero. This
can cause no modem events to ever come, including
position reports.

The fix is to not retrieve the event mask in the
client thread, but instead wait for msg to be
handled in msg task thread before retrieving it.

CRs-fixed: 2219519

Bug: 77499107
Test: reboot device and test GPS 400 times pass
Change-Id: I84495a8740b33d98ccd9c6b352742ffb0bf53c76