Revert "should be flp client if only trackingCb is registered"
This reverts commit 789e54a86ba73cfb599a9b582977b2ba7fc58e74.

CR 2912212 is needed only for location.lnx.4.0.
In 4.0, FLP tracking is handled by FlpAdapter.
The nature of the change is such that if a client is FLP client,
it shouldn’t be counted as Gnss client at the same time.
But from 5.0 onwards, all tracking (both FLP and GNSS)
are handled by GnssAdapter.
So this change would not be applicable

Change-Id: I46595b4f0783cae87c66c7e6341ceaa3a49988bc
CRs-fixed: 2952289