Z-axis support implementation
By design, in order to get location with Z,
We blindly call LBSProxy API populateAltitudeAndBroadCast.

If IzatProvider is not available then we would get
back the location without Z from LBSProxy.

If IzatProvider is available. We send
the location to Izatprovider to get the location with Z,
given Z-provider is connected.

Change-Id: Ia71019f085645a349f386bacc614f1938ee3435b
CRs-fixed: 2943684