Remove config check and caching
FW send's the LPPe/SUPL config's one at a time. In Android S,
it is seen that these config's are comming to HAL process from
HAL(Agnss) and AIDL(GnssConfiguration) api's, HAL process ends
up having two GnssAPIClient. Due to this, if on DD sub change
we end up rejecting few config update. As this transient problem
with Android S. But even if there is single GnssAPIClient object,
there is still a chance of rejection of some config on DDsub change
and leave the modem Active DD sub config in inconsistent state.
Hence removing the gnss config caching and check's for update.

Change-Id: I625c1a38f51800a6474a747213aaec7cfb7edca1
CRs-fixed: 3084143