stop using NameVals logging is about to exit
register a callback to atexit(), which would set a flag, upon which
action further use of NameVal maps will be stopped. This is because
most or all of the NameVals tbls are static unordered_maps, which
get destroyed upon exit() as well. Registering this callback here
is guaranteed that it gets called before the destroy of the static
unordered_maps.

Change-Id: Ieede9e4c09b223c31a8f249e86a32964201a6621
CRs-Fixed: 3333084