stack overflow hazard
When a large msg is being concatenated in LocIpc, a tempBuff
of 7120 bytes is allocated on the stack. This is not too far
from the 1 MB stack boundard of Android. We have a strange
issue right now in this area, with a deep stack. Overflow is
our only bet. This is a hazard nonetheless.

Change-Id: Ie4385a7c781143debb75d6f8ea343137598502be
CRs-Fixed: 3377417